inital branch commit pre merge

+++ src/Ai/Base/Actions/QuestAction.cpp
@@ -1,11 +1,12 @@
 /*
- * This file is part of the mod-playerbots module for AzerothCore. See AUTHORS file for Copyright
- * information; released under GNU GPL v2 license, redistribute/modify under version 2 of the License,
- * or (at your option) any later version.
+ * Copyright (C) 2016+ AzerothCore <www.azerothcore.org>, released under GNU AGPL v3 license, you may redistribute it
+ * and/or modify it under version 3 of the License, or (at your option), any later version.
  */
 
 #include "QuestAction.h"
-#include "BroadcastHelper.h"
+#include <sstream>
+#include <algorithm>
+
 #include "Chat.h"
 #include "ChatHelper.h"
 #include "Event.h"
@@ -15,8 +16,7 @@
 #include "Playerbots.h"
 #include "ReputationMgr.h"
 #include "ServerFacade.h"
-#include <algorithm>
-#include <sstream>
+#include "BroadcastHelper.h"
 
 bool QuestAction::Execute(Event event)
 {
@@ -413,7 +413,7 @@ bool QuestItemPushResultAction::Execute(Event event)
                 continue;
 
             int32 previousCount = itemCount - count;
-            if (itemId == itemEntry && uint32(previousCount) < quest->RequiredItemCount[i])
+            if (itemId == itemEntry && previousCount < quest->RequiredItemCount[i])
             {
                 if (botAI->GetMaster())
                 {
