+++ src/Ai/Base/Trigger/GenericTriggers.cpp
@@ -1,18 +1,17 @@
 /*
- * This file is part of the mod-playerbots module for AzerothCore. See AUTHORS file for Copyright
- * information; released under GNU GPL v2 license, redistribute/modify under version 2 of the License,
- * or (at your option) any later version.
+ * Copyright (C) 2016+ AzerothCore <www.azerothcore.org>, released under GNU AGPL v3 license, you may redistribute it
+ * and/or modify it under version 3 of the License, or (at your option), any later version.
  */
 
 #include "GenericTriggers.h"
-#include "Corpse.h"
-#include "CreatureAI.h"
+
+#include <string>
+
 #include "GenericBuffUtils.h"
+#include "CreatureAI.h"
 #include "ItemVisitors.h"
 #include "LastSpellCastValue.h"
 #include "ObjectGuid.h"
-#include "Player.h"
-#include "PlayerbotAI.h"
 #include "PlayerbotAIConfig.h"
 #include "Playerbots.h"
 #include "PositionValue.h"
@@ -20,7 +19,9 @@
 #include "TemporarySummon.h"
 #include "ThreatManager.h"
 #include "Timer.h"
-#include <string>
+#include "PlayerbotAI.h"
+#include "Player.h"
+#include "Corpse.h"
 
 bool LowManaTrigger::IsActive()
 {
@@ -163,7 +164,7 @@ bool BuffTrigger::IsActive()
         return false;
 
     Aura* aura = botAI->GetAura(spell, target, checkIsOwner, checkDuration);
-    if (!aura || (beforeDuration && uint32(aura->GetDuration()) < beforeDuration))
+    if (!aura || (beforeDuration && aura->GetDuration() < beforeDuration))
         return true;
 
     return false;
@@ -417,7 +418,7 @@ bool HealerShouldAttackTrigger::IsActive()
     return true;
 }
 
-bool ItemCountTrigger::IsActive() { return AI_VALUE2(uint32, "item count", item) < uint32(count); }
+bool ItemCountTrigger::IsActive() { return AI_VALUE2(uint32, "item count", item) < count; }
 
 bool InterruptSpellTrigger::IsActive()
 {
