+++ src/Ai/Base/Value/SpellIdValue.cpp
@@ -1,10 +1,10 @@
 /*
- * This file is part of the mod-playerbots module for AzerothCore. See AUTHORS file for Copyright
- * information; released under GNU GPL v2 license, redistribute/modify under version 2 of the License,
- * or (at your option) any later version.
+ * Copyright (C) 2016+ AzerothCore <www.azerothcore.org>, released under GNU AGPL v3 license, you may redistribute it
+ * and/or modify it under version 3 of the License, or (at your option), any later version.
  */
 
 #include "SpellIdValue.h"
+
 #include "ChatHelper.h"
 #include "Playerbots.h"
 #include "Vehicle.h"
@@ -30,7 +30,7 @@ uint32 SpellIdValue::Calculate()
 
     wstrToLower(wnamepart);
     char firstSymbol = tolower(namepart[0]);
-    size_t spellLength = wnamepart.length();
+    int spellLength = wnamepart.length();
 
     LocaleConstant loc = LOCALE_enUS;
 
@@ -133,13 +133,13 @@ uint32 SpellIdValue::Calculate()
                 continue;
             }
 
-            if (!highestRank || (uint32)id > highestRank)
+            if (!highestRank || id > highestRank)
             {
                 highestRank = id;
                 highestSpellId = spellId;
             }
 
-            if (!lowestRank || (lowestRank && (uint32)id < lowestRank))
+            if (!lowestRank || (lowestRank && id < lowestRank))
             {
                 lowestRank = id;
                 lowestSpellId = spellId;
@@ -153,7 +153,7 @@ uint32 SpellIdValue::Calculate()
             auto spellId = *it;
             if (!highestSpellId)
                 highestSpellId = spellId;
-            if (saveMana == (int32)rank)
+            if (saveMana == rank)
                 return spellId;
             lowestSpellId = spellId;
             rank++;
@@ -192,7 +192,7 @@ uint32 VehicleSpellIdValue::Calculate()
 
     wstrToLower(wnamepart);
     char firstSymbol = tolower(namepart[0]);
-    size_t spellLength = wnamepart.length();
+    int spellLength = wnamepart.length();
 
     const int loc = LocaleConstant::LOCALE_enUS;
